refactor base.html ui
