+++ src/static/base.css
@@ -1,0 +1,9 @@
+@import url('https://fonts.googleapis.com/css2?family=Lily+Script+One&display=swap');
+
+
+a#jobcrunchertitle.navbar-brand {
+    font-family: 'Lily Script One', cursive !important;
+    color: lightseagreen;
+    text-shadow: 0px 1px 1px white;
+    font-size: 30px;
+}
