+++ src/templates/base.html
@@ -1,5 +1,7 @@
 <!doctype html>
+
 <html lang="en">
+    
 <style>
     img.gif{
   width: 180px;
@@ -15,18 +17,27 @@
     <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <!-- <link href="https://fonts.googleapis.com/css2?family=Lily+Script+One&display=swap" rel="stylesheet"> -->
     <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="//use.fontawesome.com/releases/v5.15.4/css/all.css">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
         integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
 
     <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='main.css') }}">
+    <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='base.css') }}">
+    
+    <!-- <link rel="stylesheet" href="base.css"> -->
+    <!-- <link rel="stylesheet" type= "text/css" href="../styles/base.css"> -->
+
     <title>{% block title %} {% endblock %}</title>
 </head>
 
 <body>
-    <nav class="navbar navbar-expand-md navbar-light bg-light">
-        <a class="navbar-brand" href="{{ url_for('index')}}">JobCruncher</a>
+    <i class="fa-solid fa-magnifying-glass-dollar"></i>
+    <nav class="navbar navbar-expand-md navbar-dark bg-dark">
+        <a class="navbar-brand" id="jobcrunchertitle" href="{{ url_for('index')}}" >JobCruncher</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav"
             aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
@@ -34,12 +45,21 @@
         <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav">
                 <li class="nav-item active">
-                    <a class="nav-link" href="{{ url_for('search')}}">Search</a>
+                    <a class="nav-link" href="{{ url_for('index')}}"><i class="fas fa-home"></i>&nbsp; Home</a>
+                </li>
+                <li class="nav-item active">
+                    <a class="nav-link" href="{{ url_for('search')}}"><i class="fa fa-search" aria-hidden="true"></i>&nbsp; Search for Jobs</a>
                 </li>
             </ul>
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item active">
-                    <a class="nav-link" href="{{ url_for('index')}}">Back to Homepage</a>
+                    <a class="nav-link" href="{{ url_for('index')}}">Login</a>
+                </li>
+                <li class="nav-item active">
+                    <a class="nav-link" href="{{ url_for('index')}}"><i class="fas fa-user"></i>&nbsp; Profile</a>
+                </li>
+                <li class="nav-item active">
+                    <a class="nav-link" href="{{ url_for('index')}}"><i class="fas fa-sign-out-alt"></i>&nbsp; Logout</a>
                 </li>
             </ul>
         </div>
